VBA code for Automating Student File Cleaning for NSS Project
This repository contains a suite of VBA code designed to streamline and automate the data cleaning process for student files received as part of the National Student Survey. The codebase is actively maintained and will be updated iteratively to address any issues encountered during the ongoing process of file cleaning and data processing.